LGTM for libdpd (#1625)
* fix file2_init.cc

* fix file2_mat_rd.cc

* More libdpd fixes

* Finish libdpd

* Add missing header

* Add PSIEXCEPTION message in psi4/src/psi4/libdpd/pairnum.cc


* Add PSIEXCEPTION message in psi4/src/psi4/libdpd/pairnum.cc

+++ psi4/src/psi4/libdpd/file2_mat_rd.cc
@@ -52,7 +52,7 @@ int DPD::file2_mat_rd(dpdfile2 *File) {
         coltot = File->params->coltot[h ^ my_irrep];
 
         if (rowtot && coltot)
-            psio_read(File->filenum, File->label, (char *)File->matrix[h][0], rowtot * coltot * sizeof(double),
+            psio_read(File->filenum, File->label, (char *)File->matrix[h][0], static_cast<size_t>(rowtot) * coltot * sizeof(double),
                       irrep_ptr, &next_address);
     }
